refactor: move sdk partition_mapper to a standalone package
